Add files via upload
Added the ps1 file. It includes functionality to populate an array of the interface groups, and ACLs in the firepower FMC, modify the interfaces in the ACL entries, then push the updates to the rest api. It can be modified to be used for network interfaces and static routes, and modifying ACE's in an ACL as well as well.